pyproject.toml
[project]
name = "qa-bot"
version = "0.1.0"
description = "QA assistant"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cached-path>=1.6.7",
    "chromadb>=0.5.23",
    "detoxify>=0.5.2",
    "guardrails-ai>=0.5.13",
    "langchain-chroma>=0.1.4",
    "langchain-core>=0.3.28",
    "langchain-experimental>=0.3.4",
    "langchain-ollama>=0.2.3",
    "langchain>=0.3.13",
    "numpy>=1.26.4",
    "ollama>=0.4.4",
    "opik>=1.6.13",
    "pylint>=3.3.3",
    "ragas>=0.2.14",
    "sentence-transformers>=3.3.1",
    "streamlit>=1.41.1",
    "statistics>=1.0.3.5",
    "langdetect>=1.0.9",
    "loguru>=0.7.3",
    "unidecode>=1.4.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
